+++ README.md
@@ -14,3 +14,7 @@ A few resources to get you started if this is your first Flutter project:
 For help getting started with Flutter, view our
 [online documentation](https://flutter.dev/docs), which offers tutorials,
 samples, guidance on mobile development, and a full API reference.
+
+
+##TODO List
+ - 일단 카메라는 하나씩 확인가능. MLKit 확인하고 여러 사물이 한번에 인식 가능하면 기능을 변경해보자.
