+++ lib/main.dart
@@ -1,8 +1,9 @@
 import 'package:camera/camera.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_storage/firebase_storage.dart' as fireStore;
 import 'package:flutter/material.dart';
 import 'app.dart';
-
-//import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_core/firebase_core.dart';
 
 import 'app.dart';
 
@@ -10,9 +11,7 @@ List<CameraDescription> cameras;
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-
+  await Firebase.initializeApp();
   cameras = await availableCameras();
-
-  //await Firebase.initializeApp();
   runApp(MangoApp());
 }
