+++ pubspec.yaml
@@ -34,9 +34,13 @@ dependencies:
   modal_bottom_sheet: 1.0.0+1
   flutter_keyboard_visibility: ^4.0.0
   image_picker: ^0.6.7+14
-  camera: ^0.6.4+5
+  camera: ^0.5.8+2
   path: ^1.7.0
   path_provider: ^1.6.27
+  firebase_storage: ^4.0.0
+  firebase_core: ^0.5.2
+  firebase_auth: ^0.18.0+1
+  google_sign_in: "^4.5.1"
 
 dev_dependencies:
   flutter_test:
